Rename and added docs to menu
